requirements.txt
flask
python-dotenv
requests
opentelemetry-distro 
opentelemetry-exporter-jaeger 
opentelemetry-instrumentation-flask